requirements.txt
asgiref==3.3.0
Django==3.1.3
pytz==2020.4
sqlparse==0.4.1
python-decouple==3.3
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
